+++ Passepartout/App/macOS/CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## 1.15.0 Beta 2528 (2021-01-26)
+## Unreleased
 
 ### Added
 
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Provider infrastructure selectors not reloaded on profile change.
+- Missing PAC URL in proxy settings on macOS.
 
 ## 1.0.0 (2021-01-01)